ml_models/mouvements_consommation/requirements.txt
alembic==1.13.2
altair==5.4.1
aniso8601==9.0.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.32.0
Deprecated==1.2.14
docker==7.1.0
Flask==3.0.3
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==23.0.0
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mlflow==2.16.0
mlflow-skinny==2.16.0
narwhals==1.6.2
numpy==2.1.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.24.0
proto-plus==1.24.0
protobuf==5.28.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.34
sqlparse==0.5.1
st-theme==1.2.3
streamlit==1.38.0
streamlit-navigation-bar==3.3.0
streamlit-option-menu==0.3.13
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.1
